requirements.txt
openai>=1.50.0
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.2.0
pydantic>=2.9.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.14.0
